move strategy details components
