+++ apps/earn-protocol/components/layout/StrategyDetailsView/StrategyDetailsView.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import { StrategyDetailsFaq } from '@/components/organisms/StrategyDetailsFaq/StrategyDetailsFaq'
-import { StrategyDetailsHowItWorks } from '@/components/organisms/StrategyDetailsHowItWorks/StrategyDetailsHowItWorks'
-import { StrategyDetailsSecurity } from '@/components/organisms/StrategyDetailsSecurity/StrategyDetailsSecurity'
-import { StrategyDetailsYields } from '@/components/organisms/StrategyDetailsYields/StrategyDetailsYields'
+import { StrategyDetailsFaq } from '@/features/strategy-details/components/StrategyDetailsFaq/StrategyDetailsFaq'
+import { StrategyDetailsHowItWorks } from '@/features/strategy-details/components/StrategyDetailsHowItWorks/StrategyDetailsHowItWorks'
+import { StrategyDetailsSecurity } from '@/features/strategy-details/components/StrategyDetailsSecurity/StrategyDetailsSecurity'
+import { StrategyDetailsYields } from '@/features/strategy-details/components/StrategyDetailsYields/StrategyDetailsYields'
 
 export const StrategyDetailsView = () => {
   return (
