+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsFaq/StrategyDetailsFaq.tsx
