+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsSecurity/StrategyDetailsSecurity.module.scss
