+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsSecurity/StrategyDetailsSecurityMoneyExpander.tsx
