+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsYields/StrategyDetailsYields.tsx
@@ -1,18 +1,18 @@
 'use client'
 import { type FC, useEffect, useMemo, useState } from 'react'
 import { Card, InlineButtons, Table, Text, useHash } from '@summerfi/app-earn-ui'
-import { type InlineButtonOption } from '@summerfi/app-types'
+import { type InlineButtonOption } from '../../../../../../packages/app-types'
 
 import { MockedLineChart } from '@/components/organisms/Charts/MockedLineChart'
-import { individualYieldsColumns } from '@/components/organisms/StrategyDetailsYields/columns'
+import { individualYieldsColumns } from '@/features/strategy-details/components/StrategyDetailsYields/columns'
 import {
   historicalYieldOptions,
   individualYieldOptions,
   strategyDetailsYieldOptions,
   type YieldOption,
   yieldsRawData,
-} from '@/components/organisms/StrategyDetailsYields/config'
-import { individualYieldsMapper } from '@/components/organisms/StrategyDetailsYields/mapper'
+} from '@/features/strategy-details/components/StrategyDetailsYields/config'
+import { individualYieldsMapper } from '@/features/strategy-details/components/StrategyDetailsYields/mapper'
 
 export const StrategyDetailsYields: FC = () => {
   const hash = useHash<YieldOption>()
