+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsYields/columns.tsx
