+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsYields/mapper.tsx
